+++ weblate/trans/checks.py
@@ -212,7 +212,25 @@ class SameCheck(Check):
     name = _('Not translated')
     description = _('Source and translated strings are same')
 
+    def is_format_only(self, msg, regex):
+        '''
+        Checks whether given string contains only format strings
+        and possible punctation. These are quite often not changed
+        by translators.
+        '''
+        stripped = regex.sub('', msg)
+        return stripped.strip(' ,./<>?;\'\\:"|[]{}`~!@#$%^&*()-=_+') == ''
+
+
     def check_single(self, source, target, flags, language, unit):
+        # Ignore strings which don't contain any string to translate
+        if 'python-format' in flags and self.is_format_only(source, PYTHON_PRINTF_MATCH):
+            return False
+        elif 'php-format' in flags and self.is_format_only(source, PHP_PRINTF_MATCH):
+            return False
+        elif 'c-format' in flags and self.is_format_only(source, C_PRINTF_MATCH):
+            return False
+
         # One letter things are usually labels or decimal/thousand separators
         if len(source) == 1 and len(target) == 1:
             return False